# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: CivitasOne Suite API
  version: "0.2.0"
  description: |
    Unified API for CivitasOne government ERP suite.
    All endpoints require Bearer token authentication (Keycloak OIDC).
    Write operations return 202 Accepted (CQRS async processing).
  contact:
    name: CivitasOne Platform Team
    url: https://civitasone.app/docs/api
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: https://api.civitasone.app
    description: Production
  - url: http://localhost:8080
    description: Local Development
security:
  - bearerAuth: []
...
paths:
...
  /v1/helpdesk/tickets:
    post:
      operationId: createTicket
      tags: [Helpdesk]
      summary: Create a support ticket
      description: |
        Creates a new internal helpdesk ticket for IT support, facilities,
        or administrative requests. Tickets are auto-assigned based on category
        and SLA timers start immediately.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [subject, category, description]
              properties:
                subject:
                  type: string
                  maxLength: 200
                category:
                  type: string
                  enum: [it_hardware, it_software, network, facilities, admin, other]
                priority:
                  type: string
                  enum: [low, medium, high, critical]
                  default: medium
                description:
                  type: string
                  maxLength: 5000
                attachments:
                  type: array
                  items:
                    type: string
                    format: uri
                  maxItems: 5
      responses:
        "202":
          description: Ticket creation accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
        "401":
          description: Missing or invalid bearer token
      security:
        - bearerAuth: []
    get:
      operationId: listTickets
      tags: [Helpdesk]
      summary: List helpdesk tickets
      description: |
        Returns helpdesk tickets visible to the authenticated user.
        Regular users see their own tickets; helpdesk agents see assigned tickets;
        admins see all tickets.
      parameters:
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/OffsetParam"
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, waiting_on_user, resolved, closed]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, critical]
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated ticket list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        subject:
                          type: string
                        status:
                          type: string
                        priority:
                          type: string
                        category:
                          type: string
                        assignedTo:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        slaBreached:
                          type: boolean
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          description: Missing or invalid bearer token
      security:
        - bearerAuth: []

  # ═══════════════════════════════════════════════════════════════════════════
  # Workflow Service — BPMN Import/Export
  # ═══════════════════════════════════════════════════════════════════════════
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Keycloak OIDC access token (RS256). Obtain via
        POST /realms/civitasone/protocol/openid-connect/token
        with client_credentials or authorization_code grant.
    scimBearerToken:
      type: http
      scheme: bearer
      description: |
        Static bearer token for SCIM 2.0 provisioning endpoints.
        Set via SCIM_BEARER_TOKEN environment variable on the identity-service.
        Separate from Keycloak user JWTs — used only by external IdPs
        (Azure AD, Okta, Google Workspace, MeriPehchaan).
  parameters:
    LimitParam:
      name: limit
      in: query
      description: Maximum number of items to return (default 20, max 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    OffsetParam:
      name: offset
      in: query
      description: Number of items to skip for pagination
      schema:
        type: integer
        minimum: 0
        default: 0
  schemas:
...
    AcceptedResponse:
      type: object
      required: [id, status, correlationId]
      properties:
        id:
          type: string
          format: uuid
          description: ID of the created/queued resource
        status:
          type: string
          enum: [accepted, queued]
        correlationId:
          type: string
          description: Request correlation ID for tracing
    Pagination:
      type: object
      properties:
        hasMore:
          type: boolean
        pageSize:
          type: integer
        total:
          type: integer
    ValidationError:
      type: object
      required: [code, message, correlationId, retryable]
      properties:
        code:
          type: string
          examples: ["VALIDATION_FAILED"]
        message:
          type: string
        correlationId:
          type: string
        retryable:
          type: boolean
          const: false
        fieldErrors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...